Partially completed the rating functionality
